requirements.txt
discord.py==1.7.1
requests==2.7.0
python-dotenv==0.10.1
numpy==1.20.2